chore(suite): extract useDeviceCompromisedNotification
