+++ packages/suite/src/components/suite/Preloader/Preloader.tsx
@@ -26,6 +26,7 @@ import { RouterAppWithParams } from '../../../constants/suite/routes';
 import { AnalyticsConsentScreen } from '../../../views/start/AnalyticsConsentScreen';
 import { PrerequisitesGuide } from '../PrerequisitesGuide/PrerequisitesGuide';
 import { DeviceCompromised } from '../SecurityCheck/DeviceCompromised';
+import { useDeviceCompromisedNotification } from '../SecurityCheck/useDeviceCompromisedNotification';
 import { useReportDeviceCompromised } from '../SecurityCheck/useReportDeviceCompromised';
 import { LoggedOutLayout } from '../layouts/LoggedOutLayout';
 import { SuiteLayout } from '../layouts/SuiteLayout/SuiteLayout';
@@ -68,6 +69,7 @@ export const Preloader = ({ children }: PropsWithChildren) => {
     const isAnalyticsConsentConfirmed = useSelector(selectIsAnalyticsConfirmed);
 
     useReportDeviceCompromised();
+    useDeviceCompromisedNotification();
 
     const dispatch = useDispatch();
